requirements.txt
numpy
torch
transformers
codecarbon
accelerate
seaborn
matplotlib
pandas
scipy
tabulate
pydantic
scikit-learn
tensorflow
tensorflow-model-optimization
datasets
Pillow
pep8
optimum[onnxruntime]